Starting Code
The code that I have started prior to creating the repository.
Created a Game file where the main logic for the game will be and
modified the build program created by Dr.Mechtly in class that creates
the game playing area.

Creates a game playing panel and a circle for the player using buffered
images.