Smooth Player Movement
Implemented a key handler which keeps track of which keys are being
pressed or released and keeps track of the direction of movement.
Implemented a gameThread that is always running and will call the
nextStep() method every 20 milliseconds.
Implemented nextStep() method that changes the x and y coordinates
based on the booleans set for direction.