Added Enemies to Shoot
Enemy class added that creates the enemy object with its methods.
Created enemy image along with adding that to the game01.txt file
through the build class. Player has collision detection with the enemies
and the enemies can be killed when hit by a certain amount of
projectiles.  Class files also added.