Implemented Enemies Shooting Back
Modified the projectile class so there is a different projectile
creation that takes into account the players position when created so
the projectile will head that direction. Modified the enemy class to
create a mode method which will change the enemies based on which mode
it is. Modified the game class so that there HP is printed on the
player and the enemies along with creating an array list for the
enemies projectiles.